+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+# 1.2.0 - 15.12.2024
+
+- Save messages and conversations to mongoDB
+- Display markdown for both users messages and ChatGPT messages
+
 # 1.1.0 - 15.12.2024
+
 - Hided Export button
 - Restyled everything
 - Added conversation box